working utf8 string (#42)
* working utf8 string

* clean the platform layer

* clean the platform layer

* move the allocation to the platform layer

* fix linux build

* change dynamic array to memory context

* change all the containers to memory_context_t

* fix linux memory_context_t

* add arena allocator

* add memory watcher

* fix vs2017 win sdk version

* fix windows 10 sdk version

* update string documentation

* fix premake filter files

* fix windows console utf8 mode

* update the documentation

* Add visualization file for Visual Studio.

* add debug visualization to remaining containers

* Adding more containers to Vs visualizer

* Few edits on the documentations

* revert the _t convention

* Streams (#44)

* add streams

* finish file stream

* add variadic write and read

* add streams tests

* revert the scratch main file

* move concat to memory_stream

* fix linux build

+++ unittest/src/unittest_bucket_array.cpp
@@ -341,17 +341,17 @@ TEST_CASE("bucket_array test", "[bucket_array]")
 	{
 		CHECK(array.count() == 0);
 
-		for (i32 i = 0; i < 64; ++i)
+		for (usize i = 0; i < 64; ++i)
 			array.insert_back(i);
 
 		CHECK(array.count() == 64);
 
 		bucket_array<i32> array2(std::move(array));
 
-		for (i32 i = 0; i < 64; ++i)
+		for (usize i = 0; i < 64; ++i)
 			CHECK(array2[i] == i);
 
-		for (i32 i = 0; i < array.count(); ++i)
+		for (usize i = 0; i < array.count(); ++i)
 			CHECK(array[i] == i);
 
 		CHECK(array2.count() != array.count());
